nasappl: A brute-force search of 2 lastname and 2 firstname characters
This results in 456976 queries and should take ~5.3 days (~1 second per
query).  Saves the html results along the way into a /data folder
so that if the script dies it can be continued where it left off.

This script does not guarantee 100% confidence that all results will be
gathered from the NASA site but gives us a high confidence that most
results will be retreived.  The results from this script can be compared
to the original author's scrape_orig.py script to help gather confidence.

Signed-off-by: Joseph Swantek <jswantek@gogoair.com>